+++ aws_sra_examples/solutions/guardduty/guardduty_org/lambda/src/guardduty.py
@@ -11,6 +11,7 @@ from __future__ import annotations
 
 import json
 import logging
+import math
 from time import sleep
 from typing import TYPE_CHECKING, Any, Dict
 
@@ -30,10 +31,13 @@ LOGGER = logging.getLogger("sra")
 
 # Global variables
 SERVICE_NAME = "guardduty.amazonaws.com"
+PRINCIPAL_NAME = "malware-protection.guardduty.amazonaws.com"
 SLEEP_SECONDS = 10
 UNEXPECTED = "Unexpected!"
 MAX_RETRY = 5
 BOTO3_CONFIG = Config(retries={"max_attempts": 10, "mode": "standard"})
+CHECK_ACCT_MEMBER_RETRIES = 10
+
 
 try:
     MANAGEMENT_ACCOUNT_SESSION = boto3.Session()
@@ -113,6 +117,44 @@ def get_unprocessed_account_details(create_members_response: CreateMembersRespon
     return remaining_accounts
 
 
+def check_members(guardduty_client: GuardDutyClient, detector_id: str, accounts: list) -> list:
+    """Check all accounts in the organization are member accounts.
+
+    Args:
+        guardduty_client: boto3 guardduty client
+        detector_id: detectorId of the delegated admin account
+        accounts: list of accounts in the organization
+
+    Returns:
+        any account in the organization that isn't a member
+    """
+    LOGGER.info("check_members begin")
+    retries = 0
+    missing_members: list = []
+    confirmed_members: list = []
+    while retries < CHECK_ACCT_MEMBER_RETRIES:
+        confirmed_members = []
+        missing_members = []
+        member_paginator = guardduty_client.get_paginator("list_members")
+        page_iterator = member_paginator.paginate(DetectorId=detector_id)
+        for page in page_iterator:
+            for member in page["Members"]:
+                confirmed_members.append(member["AccountId"])
+        for account in accounts:
+            if account["AccountId"] not in confirmed_members:
+                missing_members.append(account)
+        if len(missing_members) > 0:
+            LOGGER.info(f"missing {len(missing_members)} members: {missing_members}")
+            retries += 1
+            LOGGER.info(f"sleep for {SLEEP_SECONDS} retry number {retries}")
+            sleep(SLEEP_SECONDS)
+        else:
+            LOGGER.info("All accounts in the organization are members")
+            break
+    LOGGER.info("check_members end")
+    return missing_members
+
+
 def create_members(guardduty_client: GuardDutyClient, detector_id: str, accounts: list) -> None:  # noqa: CCR001 (cognitive complexity)
     """Create GuardDuty members with existing accounts. Retry 2 times.
 
@@ -125,32 +167,59 @@ def create_members(guardduty_client: GuardDutyClient, detector_id: str, accounts
         ValueError: Unprocessed member accounts.
     """
     LOGGER.info("Creating members")
-    create_members_response = guardduty_client.create_members(DetectorId=detector_id, AccountDetails=accounts)
 
-    if "UnprocessedAccounts" in create_members_response and create_members_response["UnprocessedAccounts"]:
-        unprocessed = True
-        retry_count = 0
-        unprocessed_accounts = []
-        while unprocessed:
-            retry_count += 1
-            LOGGER.info(f"Unprocessed Accounts: {create_members_response['UnprocessedAccounts']}")
-            remaining_accounts = get_unprocessed_account_details(create_members_response, accounts)
+    number_of_create_members_calls = math.ceil(len(accounts) / 50)
 
-            unprocessed = False
-            if remaining_accounts:
-                create_members_response = guardduty_client.create_members(DetectorId=detector_id, AccountDetails=remaining_accounts)
-                if "UnprocessedAccounts" in create_members_response and create_members_response["UnprocessedAccounts"]:
-                    unprocessed_accounts = create_members_response["UnprocessedAccounts"]
-                    if retry_count != MAX_RETRY:
-                        unprocessed = True
+    for api_call_number in range(0, number_of_create_members_calls):
+        account_details = accounts[api_call_number * 50 : (api_call_number * 50) + 50]
+        LOGGER.info(f"Calling create_member, api_call_number {api_call_number} with detector_id: {detector_id}")
+        LOGGER.info(f"Create member account_details: {account_details}, account_details length: {len(account_details)}")
+        create_members_response = guardduty_client.create_members(DetectorId=detector_id, AccountDetails=account_details)
 
-        if unprocessed_accounts:
-            LOGGER.info(f"Unprocessed Member Accounts: {unprocessed_accounts}")
-            raise ValueError("Unprocessed Member Accounts")
+        if "UnprocessedAccounts" in create_members_response and create_members_response["UnprocessedAccounts"]:
+            unprocessed = True
+            retry_count = 0
+            unprocessed_accounts = []
+            LOGGER.info(f"Retry number; {retry_count} for unprocessed accounts")
+            LOGGER.info(f"Sleeping for {SLEEP_SECONDS} before retry")
+            sleep(SLEEP_SECONDS)
+            while unprocessed:
+                retry_count += 1
+                LOGGER.info(f"Unprocessed Accounts: {create_members_response['UnprocessedAccounts']}")
+                remaining_accounts = get_unprocessed_account_details(create_members_response, accounts)
+
+                if len(remaining_accounts) > 0:
+                    LOGGER.info("Remaining accounts found during create members")
+                    LOGGER.info(f"Calling create_member, api_call_number {api_call_number} with detector_id: {detector_id}")
+                    LOGGER.info(f"Create member account_details: {remaining_accounts}, remaining_accounts length: {len(remaining_accounts)}")
+                    create_members_response = guardduty_client.create_members(DetectorId=detector_id, AccountDetails=remaining_accounts)
+                    if "UnprocessedAccounts" in create_members_response and create_members_response["UnprocessedAccounts"]:
+                        LOGGER.info("Unprocessed accounts found during retry")
+                        unprocessed_accounts = create_members_response["UnprocessedAccounts"]
+                        if retry_count == MAX_RETRY:
+                            unprocessed = False
+
+            if unprocessed_accounts:
+                LOGGER.info(f"Unprocessed Member Accounts: {unprocessed_accounts}")
+                raise ValueError("Unprocessed Member Accounts while Creating Members")
+
+    missing_members: list = check_members(guardduty_client, detector_id, accounts)
+    if len(missing_members) > 0:
+        LOGGER.info(f"Check members failure: {missing_members}")
+        raise ValueError("Check members failure")
 
 
-def update_member_detectors(  # noqa: CCR001 (cognitive complexity)
-    guardduty_client: GuardDutyClient, detector_id: str, account_ids: list, auto_enable_s3_logs: bool
+def update_member_detectors(  # noqa: CCR001, C901, CFQ002, CFQ001
+    guardduty_client: GuardDutyClient,
+    detector_id: str,
+    account_ids: list,
+    auto_enable_s3_logs: bool,
+    enable_eks_audit_logs: bool,
+    auto_enable_malware_protection: bool,
+    enable_rds_login_events: bool,
+    enable_eks_runtime_monitoring: bool,
+    enable_eks_addon_management: bool,
+    enable_lambda_network_logs: bool,
 ) -> None:
     """Update member detectors.
 
@@ -158,16 +227,103 @@ def update_member_detectors(  # noqa: CCR001 (cognitive complexity)
         guardduty_client: GuardDuty client
         detector_id: GuardDuty detector id
         account_ids: member account list
-        auto_enable_s3_logs: Auto Enable S3 Logs
+        auto_enable_s3_logs: Auto enable S3 Logs
+        enable_eks_audit_logs: Auto enable Kubernetes Audit Logs
+        auto_enable_malware_protection: Auto enable Malware Protection
+        enable_rds_login_events: Auto enable RDS login activity monitoring
+        enable_eks_runtime_monitoring: Auto enable EKS runtime monitoring
+        enable_eks_addon_management: Auto enable EKS add-on
+        enable_lambda_network_logs: Auto enable Lambda network logs
 
     Raises:
         ValueError: Unprocessed member accounts
     """
-    configuration_params: Dict[str, Any] = {"DetectorId": detector_id, "AccountIds": account_ids}
+    LOGGER.info("Updating Member Detectors")
+    configuration_params: Dict[str, Any] = {"DetectorId": detector_id, "AccountIds": account_ids, "Features": []}
+    number_of_create_members_calls: int = math.ceil(len(configuration_params["AccountIds"]) / 50)
+    LOGGER.info("Setting feature configuration parameters once...")
     if auto_enable_s3_logs:
-        configuration_params["DataSources"] = {"S3Logs": {"Enable": True}}
+        configuration_params["Features"].append({"Name": "S3_DATA_EVENTS", "Status": "ENABLED"})
+    else:
+        configuration_params["Features"].append({"Name": "S3_DATA_EVENTS", "Status": "DISABLED"})
+    if enable_eks_audit_logs:
+        configuration_params["Features"].append({"Name": "EKS_AUDIT_LOGS", "Status": "ENABLED"})
+    else:
+        configuration_params["Features"].append({"Name": "EKS_AUDIT_LOGS", "Status": "DISABLED"})
+    if auto_enable_malware_protection:
+        configuration_params["Features"].append({"Name": "EBS_MALWARE_PROTECTION", "Status": "ENABLED"})
+    else:
+        configuration_params["Features"].append({"Name": "EBS_MALWARE_PROTECTION", "Status": "DISABLED"})
+    if enable_rds_login_events:
+        configuration_params["Features"].append({"Name": "RDS_LOGIN_EVENTS", "Status": "ENABLED"})
+    else:
+        configuration_params["Features"].append({"Name": "RDS_LOGIN_EVENTS", "Status": "DISABLED"})
+    if enable_eks_runtime_monitoring and enable_eks_addon_management:
+        configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "Status": "ENABLED",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "Status": "ENABLED"}],
+            }
+        )
+    elif enable_eks_runtime_monitoring and not enable_eks_addon_management:
+        configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "Status": "ENABLED",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "Status": "DISABLED"}],
+            }
+        )
+    else:
+        configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "Status": "DISABLED",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "Status": "DISABLED"}],
+            }
+        )
+    if enable_lambda_network_logs:
+        configuration_params["Features"].append({"Name": "LAMBDA_NETWORK_LOGS", "Status": "ENABLED"})
+    else:
+        configuration_params["Features"].append({"Name": "LAMBDA_NETWORK_LOGS", "Status": "DISABLED"})
 
-    update_member_response = guardduty_client.update_member_detectors(**configuration_params)
+    LOGGER.info("Iterating through api calls for each group of accounts...")
+    for api_call_number in range(0, number_of_create_members_calls):
+        configuration_params["AccountIds"] = account_ids[api_call_number * 50 : (api_call_number * 50) + 50]
+        LOGGER.info(f"Configuration parameters are: {configuration_params}")
+
+        LOGGER.info(f"Calling update_member_detectors with params {configuration_params}")
+        update_member_response = guardduty_client.update_member_detectors(**configuration_params)
+
+        if "UnprocessedAccounts" in update_member_response and update_member_response["UnprocessedAccounts"]:
+            unprocessed = True
+            retry_count = 0
+            unprocessed_accounts = []
+            while unprocessed:
+                LOGGER.info(f"Unprocessed accounts found. Retry number; {retry_count} for unprocessed accounts")
+                LOGGER.info(f"Sleeping for {SLEEP_SECONDS} before retry")
+                sleep(SLEEP_SECONDS)
+                retry_count += 1
+                remaining_accounts = []
+
+                for unprocessed_account in update_member_response["UnprocessedAccounts"]:
+                    if unprocessed_account["AccountId"] in account_ids:
+                        remaining_accounts.append(unprocessed_account["AccountId"])
+
+                if len(remaining_accounts) > 0:
+                    configuration_params["AccountIds"] = remaining_accounts
+                    LOGGER.info(f"Remaining accounts found during update_member_detectors {remaining_accounts}")
+                    LOGGER.info(f"Calling retry update_member_detectors with params {configuration_params}")
+                    update_member_response = guardduty_client.update_member_detectors(**configuration_params)
+                    if "UnprocessedAccounts" in update_member_response and update_member_response["UnprocessedAccounts"]:
+                        LOGGER.info(f"Unprocessed accounts found during retry: {update_member_response['UnprocessedAccounts']}")
+                        LOGGER.info(f"Calling update_member_detectors with params {configuration_params}")
+                        unprocessed_accounts = update_member_response["UnprocessedAccounts"]
+                        if retry_count == 5:
+                            LOGGER.info("retry count is 5 setting unprocessed to false")
+                            unprocessed = False
+                    else:
+                        LOGGER.info("No more unprocessed accounts found setting unprocessed to false")
 
     if "UnprocessedAccounts" in update_member_response and update_member_response["UnprocessedAccounts"]:
         unprocessed = True
@@ -189,42 +345,163 @@ def update_member_detectors(  # noqa: CCR001 (cognitive complexity)
                     unprocessed_accounts = update_member_response["UnprocessedAccounts"]
                     if retry_count == 5:
                         unprocessed = False
-                else:
-                    unprocessed = False
 
-        if unprocessed_accounts:
-            LOGGER.info(f"Update Member Detectors Unprocessed Member Accounts: {unprocessed_accounts}")
-            raise ValueError("Unprocessed Member Accounts")
+            if unprocessed_accounts:
+                LOGGER.info(f"Update Member Detectors Unprocessed Member Accounts: {unprocessed_accounts}")
+                raise ValueError("Unprocessed Member Accounts while Updating Member Detectors")
 
 
-def update_guardduty_configuration(
-    guardduty_client: GuardDutyClient, auto_enable_s3_logs: bool, detector_id: str, finding_publishing_frequency: str, account_ids: list
+def update_guardduty_configuration(  # noqa: CCR001, C901, CFQ002, CFQ001
+    guardduty_client: GuardDutyClient,
+    auto_enable_s3_logs: bool,
+    enable_eks_audit_logs: bool,
+    auto_enable_malware_protection: bool,
+    enable_rds_login_events: bool,
+    enable_eks_runtime_monitoring: bool,
+    enable_eks_addon_management: bool,
+    enable_lambda_network_logs: bool,
+    detector_id: str,
+    finding_publishing_frequency: str,
+    account_ids: list,
 ) -> None:
     """Update GuardDuty configuration to auto enable new accounts and S3 log protection.
 
     Args:
         guardduty_client: GuardDuty Client
         auto_enable_s3_logs: Auto enable S3 Logs
+        enable_eks_audit_logs: Auto enable Kubernetes Audit Logs
+        auto_enable_malware_protection: Auto enable Malware Protection
+        enable_rds_login_events: Auto enable RDS login activity monitoring
+        enable_eks_runtime_monitoring: Auto enable EKS runtime monitoring
+        enable_eks_addon_management: Auto enable EKS add-on
+        enable_lambda_network_logs: Auto enable Lambda network logs
         detector_id: GuardDuty detector ID
         finding_publishing_frequency: Finding publishing frequency
         account_ids: List of member account ids
     """
-    org_configuration_params: Dict[str, Any] = {"DetectorId": detector_id, "AutoEnable": True}
-    admin_configuration_params: Dict[str, Any] = {"DetectorId": detector_id, "FindingPublishingFrequency": finding_publishing_frequency}
+    org_configuration_params: Dict[str, Any] = {"DetectorId": detector_id, "AutoEnable": True, "Features": []}
+    admin_configuration_params: Dict[str, Any] = {
+        "DetectorId": detector_id,
+        "FindingPublishingFrequency": finding_publishing_frequency,
+        "Features": [],
+    }
 
     if auto_enable_s3_logs:
-        org_configuration_params["DataSources"] = {"S3Logs": {"AutoEnable": True}}
-        admin_configuration_params["DataSources"] = {"S3Logs": {"Enable": True}}
+        org_configuration_params["Features"].append({"Name": "S3_DATA_EVENTS", "AutoEnable": "NEW"})
+    else:
+        org_configuration_params["Features"].append({"Name": "S3_DATA_EVENTS", "AutoEnable": "NONE"})
+    if enable_eks_audit_logs:
+        org_configuration_params["Features"].append({"Name": "EKS_AUDIT_LOGS", "AutoEnable": "NEW"})
+    else:
+        org_configuration_params["Features"].append({"Name": "EKS_AUDIT_LOGS", "AutoEnable": "NONE"})
+    if auto_enable_malware_protection:
+        org_configuration_params["Features"].append({"Name": "EBS_MALWARE_PROTECTION", "AutoEnable": "NEW"})
+    else:
+        org_configuration_params["Features"].append({"Name": "EBS_MALWARE_PROTECTION", "AutoEnable": "NONE"})
+    if enable_rds_login_events:
+        org_configuration_params["Features"].append({"Name": "RDS_LOGIN_EVENTS", "AutoEnable": "NEW"})
+    else:
+        org_configuration_params["Features"].append({"Name": "RDS_LOGIN_EVENTS", "AutoEnable": "NONE"})
+    if enable_eks_runtime_monitoring and enable_eks_addon_management:
+        org_configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "AutoEnable": "NEW",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "AutoEnable": "NEW"}],
+            }
+        )
+    elif enable_eks_runtime_monitoring and not enable_eks_addon_management:
+        org_configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "AutoEnable": "NEW",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "AutoEnable": "NONE"}],
+            }
+        )
+    else:
+        org_configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "AutoEnable": "NONE",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "AutoEnable": "NONE"}],
+            }
+        )
+    if enable_lambda_network_logs:
+        org_configuration_params["Features"].append({"Name": "LAMBDA_NETWORK_LOGS", "AutoEnable": "NEW"})
+    else:
+        org_configuration_params["Features"].append({"Name": "LAMBDA_NETWORK_LOGS", "AutoEnable": "NONE"})
+
+    if auto_enable_s3_logs:
+        admin_configuration_params["Features"].append({"Name": "S3_DATA_EVENTS", "Status": "ENABLED"})
+    else:
+        admin_configuration_params["Features"].append({"Name": "S3_DATA_EVENTS", "Status": "DISABLED"})
+    if enable_eks_audit_logs:
+        admin_configuration_params["Features"].append({"Name": "EKS_AUDIT_LOGS", "Status": "ENABLED"})
+    else:
+        admin_configuration_params["Features"].append({"Name": "EKS_AUDIT_LOGS", "Status": "DISABLED"})
+    if auto_enable_malware_protection:
+        admin_configuration_params["Features"].append({"Name": "EBS_MALWARE_PROTECTION", "Status": "ENABLED"})
+    else:
+        admin_configuration_params["Features"].append({"Name": "EBS_MALWARE_PROTECTION", "Status": "DISABLED"})
+    if enable_rds_login_events:
+        admin_configuration_params["Features"].append({"Name": "RDS_LOGIN_EVENTS", "Status": "ENABLED"})
+    else:
+        admin_configuration_params["Features"].append({"Name": "RDS_LOGIN_EVENTS", "Status": "DISABLED"})
+    if enable_eks_runtime_monitoring and enable_eks_addon_management:
+        admin_configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "Status": "ENABLED",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "Status": "ENABLED"}],
+            }
+        )
+    elif enable_eks_runtime_monitoring and not enable_eks_addon_management:
+        admin_configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "Status": "ENABLED",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "Status": "DISABLED"}],
+            }
+        )
+    else:
+        admin_configuration_params["Features"].append(
+            {
+                "Name": "EKS_RUNTIME_MONITORING",
+                "Status": "DISABLED",
+                "AdditionalConfiguration": [{"Name": "EKS_ADDON_MANAGEMENT", "Status": "DISABLED"}],
+            }
+        )
+    if enable_lambda_network_logs:
+        admin_configuration_params["Features"].append({"Name": "LAMBDA_NETWORK_LOGS", "Status": "ENABLED"})
+    else:
+        admin_configuration_params["Features"].append({"Name": "LAMBDA_NETWORK_LOGS", "Status": "DISABLED"})
 
     guardduty_client.update_organization_configuration(**org_configuration_params)
     guardduty_client.update_detector(**admin_configuration_params)
-    update_member_detectors(guardduty_client, detector_id, account_ids, auto_enable_s3_logs)
+    update_member_detectors(
+        guardduty_client,
+        detector_id,
+        account_ids,
+        auto_enable_s3_logs,
+        enable_eks_audit_logs,
+        auto_enable_malware_protection,
+        enable_rds_login_events,
+        enable_eks_runtime_monitoring,
+        enable_eks_addon_management,
+        enable_lambda_network_logs,
+    )
 
 
-def configure_guardduty(
+def configure_guardduty(  # noqa: CFQ002, CFQ001
     session: boto3.Session,
     delegated_account_id: str,
     auto_enable_s3_logs: bool,
+    enable_eks_audit_logs: bool,
+    auto_enable_malware_protection: bool,
+    enable_rds_login_events: bool,
+    enable_eks_runtime_monitoring: bool,
+    enable_eks_addon_management: bool,
+    enable_lambda_network_logs: bool,
     region_list: list,
     finding_publishing_frequency: str,
     kms_key_arn: str,
@@ -235,7 +512,13 @@ def configure_guardduty(
     Args:
         session: boto3 session
         delegated_account_id: Delegated Admin Account ID
-        auto_enable_s3_logs: Auto Enable S3 Logs
+        auto_enable_s3_logs: Auto enable S3 Logs
+        enable_eks_audit_logs: Auto enable Kubernetes Audit Logs
+        auto_enable_malware_protection: Auto enable Malware Protection
+        enable_rds_login_events: Auto enable RDS login activity monitoring
+        enable_eks_runtime_monitoring: Auto enable EKS runtime monitoring
+        enable_eks_addon_management: Auto enable EKS add-on
+        enable_lambda_network_logs: Auto enable Lambda network logs
         region_list: AWS Regions
         finding_publishing_frequency: Finding publishing frequency
         kms_key_arn: KMS Key ARN
@@ -284,7 +567,19 @@ def configure_guardduty(
 
             LOGGER.info(f"Waiting {SLEEP_SECONDS} seconds before updating the configuration.")
             sleep(SLEEP_SECONDS)
-            update_guardduty_configuration(regional_guardduty, auto_enable_s3_logs, detector_id, finding_publishing_frequency, account_ids)
+            update_guardduty_configuration(
+                regional_guardduty,
+                auto_enable_s3_logs,
+                enable_eks_audit_logs,
+                auto_enable_malware_protection,
+                enable_rds_login_events,
+                enable_eks_runtime_monitoring,
+                enable_eks_addon_management,
+                enable_lambda_network_logs,
+                detector_id,
+                finding_publishing_frequency,
+                account_ids,
+            )
 
 
 def check_for_detectors(session: boto3.Session, regions: list) -> bool:  # noqa: CCR001 (cognitive complexity)
@@ -349,6 +644,20 @@ def process_delete_event(params: dict, regions: list, account_ids: list, include
             management_sns_client.publish(TopicArn=params["SNS_TOPIC_ARN"], Message=json.dumps(sns_message))
 
 
+def disable_aws_service_access(service_principal: str = PRINCIPAL_NAME) -> None:
+    """Disable service access for the provided service principal within AWS Organizations.
+
+    Args:
+        service_principal: Service Principal
+    """
+    try:
+        LOGGER.info(f"Disabling service access for {service_principal}")
+
+        ORG_CLIENT.disable_aws_service_access(ServicePrincipal=service_principal)
+    except ORG_CLIENT.exceptions.AccountNotRegisteredException as error:
+        LOGGER.info(f"Service ({service_principal}) does not have organizations access revoked: {error}")
+
+
 def cleanup_member_account(account_id: str, delete_detector_role_name: str, regions: list) -> dict:
     """Cleanup member account.
 
@@ -378,6 +687,7 @@ def delete_detectors(guardduty_client: GuardDutyClient, region: str, is_delegate
         region: AWS Region
         is_delegated_admin: True or False
     """
+    disable_aws_service_access(PRINCIPAL_NAME)
     detectors = guardduty_client.list_detectors()
 
     if detectors["DetectorIds"]: